+++ src/arch/cc/codegen_cc.h
@@ -7,13 +7,14 @@
  *          functions is correct
  */
 
-#ifndef QL_ARCH_CC_CODEGEN_CC_H
-#define QL_ARCH_CC_CODEGEN_CC_H
+#ifndef ARCH_CC_CODEGEN_CC_H
+#define ARCH_CC_CODEGEN_CC_H
 
 // options
 #define OPT_SUPPORT_STATIC_CODEWORDS    1
 #define OPT_VCD_OUTPUT                  1   // output Value Change Dump file for GTKWave viewer
 #define OPT_RUN_ONCE                    0   // 0=loop indefinitely (CC-light emulation)
+#define OPT_CALCULATE_LATENCIES         0   // fixed compensation based on instrument latencies
 
 #include "json.h"
 #include "platform.h"
@@ -39,7 +40,7 @@ private: // types
 
     typedef struct {
         std::string signalValue;
-        size_t duration_ns;
+        size_t durationInNs;
         ssize_t readoutCop;     // NB: we use ssize_t iso size_t so we can encode 'unused' (-1)
 #if OPT_SUPPORT_STATIC_CODEWORDS
         int staticCodewordOverride;
@@ -51,7 +52,44 @@ private: // types
         std::string path;
     } tJsonNodeInfo;
 
-private: // vars
+public:
+    codegen_cc() = default;
+    ~codegen_cc() = default;
+
+    // Generic
+    void init(const ql::quantum_platform &platform);
+    std::string getCode();
+    std::string getMap();
+
+    void program_start(const std::string &progName);
+    void program_finish(const std::string &progName);
+    void kernel_start();
+    void kernel_finish(const std::string &kernelName, size_t durationInCycles);
+    void bundle_start(const std::string &cmnt);
+    void bundle_finish(size_t startCycle, size_t durationInCycles, bool isLastBundle);
+    void comment(const std::string &c);
+
+    // Quantum instructions
+    void custom_gate(
+            const std::string &iname,
+            const std::vector<size_t> &qops,
+            const std::vector<size_t> &cops,
+            double angle, size_t startCycle, size_t durationInNs);
+    void nop_gate();
+
+    // Classical operations on kernels
+    void if_start(size_t op0, const std::string &opName, size_t op1);
+    void else_start(size_t op0, const std::string &opName, size_t op1);
+    void for_start(const std::string &label, int iterations);
+    void for_end(const std::string &label);
+    void do_while_start(const std::string &label);
+    void do_while_end(const std::string &label, size_t op0, const std::string &opName, size_t op1);
+
+    // Classical arithmetic instructions
+    void add();
+    // FIXME: etc
+
+private:    // vars
     static const int MAX_SLOTS = 12;
     static const int MAX_GROUPS = 32;                           // enough for VSM
 
@@ -66,12 +104,11 @@ private: // vars
     json inputLutTable;                                         // input LUT usage per instrument group
     size_t lastStartCycle[MAX_SLOTS];
 
-    // some JSON nodes we need access to. FIXME: use pointers/const json & for efficiency?
-    json jsonBackendSettings;
-    json jsonInstrumentDefinitions;
-    json jsonControlModes;
-    json jsonInstruments;
-    json jsonSignals;
+    // some JSON nodes we need access to
+    const json *jsonInstrumentDefinitions;
+    const json *jsonControlModes;
+    const json *jsonInstruments;
+    const json *jsonSignals;
 
     const ql::quantum_platform *platform;
 
@@ -84,50 +121,17 @@ private: // vars
     std::vector<int> vcdVarCodeword;
 #endif
 
-public:
-    codegen_cc() {}
-    ~codegen_cc() {}
-
-    // Generic
-    void init(const ql::quantum_platform &platform);
-    std::string getCode();
-    std::string getMap();
-
-    void program_start(std::string prog_name);
-    void program_finish(std::string prog_name);
-    void kernel_start();
-    void kernel_finish(std::string kernelName, size_t duration_in_cycles);
-    void bundle_start(std::string cmnt);
-    void bundle_finish(size_t start_cycle, size_t duration_in_cycles, bool isLastBundle);
-    void comment(std::string c);
-
-    // Quantum instructions
-    void custom_gate(std::string iname, std::vector<size_t> qops, std::vector<size_t> cops, double angle, size_t start_cycle, size_t duration_ns);
-    void nop_gate();
-
-    // Classical operations on kernels
-    void if_start(size_t op0, std::string opName, size_t op1);
-    void else_start(size_t op0, std::string opName, size_t op1);
-    void for_start(std::string label, int iterations);
-    void for_end(std::string label);
-    void do_while_start(std::string label);
-    void do_while_end(std::string label, size_t op0, std::string opName, size_t op1);
-
-    // Classical arithmetic instructions
-    void add();
-    // FIXME: etc
-
-private:
+private:    // funcs
     // Some helpers to ease nice assembly formatting
     void emit(const char *labelOrComment, const char *instr="");
 
     // @param   label       must include trailing ":"
     // @param   comment     must include leading "#"
-    void emit(const char *label, const char *instr, std::string qops, const char *comment="");
+    void emit(const char *label, const char *instr, const std::string &qops, const char *comment="");
 
     // helpers
     void latencyCompensation();
-    void padToCycle(size_t lastStartCycle, size_t start_cycle, int slot, std::string instrumentName);
+    void padToCycle(size_t lastStartCycle, size_t startCycle, int slot, const std::string &instrumentName);
     uint32_t assignCodeword(const std::string &instrumentName, int instrIdx, int group);
 
     // Functions processing JSON
@@ -135,10 +139,9 @@ private:
     const json &findInstrumentDefinition(const std::string &name);
 
     // find instrument/group providing instructionSignalType for qubit
-    tSignalInfo findSignalInfoForQubit(std::string instructionSignalType, size_t qubit);
+    tSignalInfo findSignalInfoForQubit(const std::string &instructionSignalType, size_t qubit);
 
     tJsonNodeInfo findSignalDefinition(const json &instruction, const std::string &iname) const;
 }; // class
 
-
-#endif  // ndef QL_ARCH_CC_CODEGEN_CC_H
+#endif  // ndef ARCH_CC_CODEGEN_CC_H